+++ lib/bullet/detector/n_plus_one_query.rb
@@ -7,7 +7,7 @@ module Bullet
       extend StackTraceFilter
 
       class << self
-        # executed when object.assocations is called.
+        # executed when object.associations is called.
         # first, it keeps this method call for object.association.
         # then, it checks if this associations call is unpreload.
         #   if it is, keeps this unpreload associations and caller.
@@ -33,8 +33,9 @@ module Bullet
           return unless Bullet.start?
           return unless Bullet.n_plus_one_query_enable?
 
-          objects = Array(object_or_objects)
+          objects = Array.wrap(object_or_objects)
           return if objects.map(&:bullet_primary_key_value).compact.empty?
+          return if objects.all? { |obj| obj.class.name =~ /^HABTM_/ }
 
           Bullet.debug(
             'Detector::NPlusOneQuery#add_possible_objects',
@@ -94,7 +95,7 @@ module Bullet
         private
 
         def create_notification(callers, klazz, associations)
-          notify_associations = Array(associations) - Bullet.get_whitelist_associations(:n_plus_one_query, klazz)
+          notify_associations = Array.wrap(associations) - Bullet.get_safelist_associations(:n_plus_one_query, klazz)
 
           if notify_associations.present?
             notice = Bullet::Notification::NPlusOneQuery.new(callers, klazz, notify_associations)