+++ lib/bullet/mongoid7x.rb
@@ -4,35 +4,50 @@ module Bullet
   module Mongoid
     def self.enable
       require 'mongoid'
+      require 'rubygems'
       ::Mongoid::Contextual::Mongo.class_eval do
         alias_method :origin_first, :first
         alias_method :origin_last, :last
         alias_method :origin_each, :each
         alias_method :origin_eager_load, :eager_load
 
-        def first(opts = {})
-          result = origin_first(opts)
-          Bullet::Detector::NPlusOneQuery.add_impossible_object(result) if result
-          result
-        end
-
-        def last(opts = {})
-          result = origin_last(opts)
-          Bullet::Detector::NPlusOneQuery.add_impossible_object(result) if result
-          result
+        %i[first last].each do |context|
+          default = Gem::Version.new(::Mongoid::VERSION) >= Gem::Version.new('7.5') ? nil : {}
+          define_method(context) do |opts = default|
+            result = send(:"origin_#{context}", opts)
+            Bullet::Detector::NPlusOneQuery.add_impossible_object(result) if result
+            result
+          end
         end
 
         def each(&block)
-          return to_enum unless block
+          return to_enum unless block_given?
 
-          records = []
-          origin_each { |record| records << record }
-          if records.length > 1
-            Bullet::Detector::NPlusOneQuery.add_possible_objects(records)
-          elsif records.size == 1
-            Bullet::Detector::NPlusOneQuery.add_impossible_object(records.first)
+          first_document = nil
+          document_count = 0
+
+          origin_each do |document|
+            document_count += 1
+
+            if document_count == 1
+              first_document = document
+            elsif document_count == 2
+              Bullet::Detector::NPlusOneQuery.add_possible_objects([first_document, document])
+              yield(first_document)
+              first_document = nil
+              yield(document)
+            else
+              Bullet::Detector::NPlusOneQuery.add_possible_objects(document)
+              yield(document)
+            end
           end
-          records.each(&block)
+
+          if document_count == 1
+            Bullet::Detector::NPlusOneQuery.add_impossible_object(first_document)
+            yield(first_document)
+          end
+
+          self
         end
 
         def eager_load(docs)