+++ spec/models/deal.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+class Deal < ActiveRecord::Base
+  has_and_belongs_to_many :posts
+end