+++ spec/support/sqlite_seed.rb
@@ -21,6 +21,13 @@ module Support
       post2 = category2.posts.create(name: 'second', writer: writer2)
       post3 = category2.posts.create(name: 'third', writer: writer2)
 
+      deal1 = Deal.new(name: 'Deal 1')
+      deal1.posts << post1
+      deal1.posts << post2
+      deal2 = Deal.new(name: 'Deal 2')
+      post1.deals << deal1
+      post1.deals << deal2
+
       comment1 = post1.comments.create(name: 'first', author: writer1)
       comment2 = post1.comments.create(name: 'first2', author: writer1)
       comment3 = post1.comments.create(name: 'first3', author: writer1)
@@ -85,8 +92,15 @@ module Support
       page3 = Page.create(name: 'page3', parent_id: folder2.id, author_id: author2.id)
       page4 = Page.create(name: 'page4', parent_id: folder2.id, author_id: author2.id)
 
+      role1 = Role.create(name: 'Admin')
+      role2 = Role.create(name: 'User')
+
       user1 = User.create(name: 'user1', category: category1)
       user2 = User.create(name: 'user2', category: category1)
+
+      user1.roles << role1
+      user1.roles << role2
+      user2.roles << role2
 
       submission1 = user1.create_submission(name: 'submission1')
       submission2 = user2.create_submission(name: 'submission2')
@@ -154,6 +168,11 @@ module Support
         create_table :deals do |t|
           t.column :name, :string
           t.column :hotel_id, :integer
+        end
+
+        create_table :deals_posts do |t|
+          t.column :deal_id, :integer
+          t.column :post_id, :integer
         end
 
         create_table :documents do |t|
@@ -232,6 +251,17 @@ module Support
         create_table :replies do |t|
           t.column :name, :string
           t.column :submission_id, :integer
+        end
+
+        create_table :roles do |t|
+          t.column :name, :string
+          t.column :resource_id, :integer
+          t.column :resource_type, :string
+        end
+
+        create_table :roles_users do |t|
+          t.column :role_id, :integer
+          t.column :user_id, :integer
         end
 
         create_table :submissions do |t|